% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{BiocMaintainerShiny}
\alias{BiocMaintainerShiny}
\title{Launch Shiny application for Bioconductor Maintainer Validation App}
\usage{
BiocMaintainerShiny(...)
}
\arguments{
\item{...}{Further arguments to the \code{runApp} function}
}
\value{
Displays shiny app in browser
}
\description{
This shiny app allows exploration of package maintainers and
validity of emails.
}
\details{
This shiny app allows exploration of package maintainers and
validity of emails. The associated Bioconductor validation app sends a yearly
reminder of Bioconductor policies that maintainers should opt-in. Th app also
reflects the status of accepting policies and whether or not the email was
reachable.
}
\examples{
if (interactive()) {
    BiocMaintainerShiny()
}
}
\author{
Lori Shepherd
}
