% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiWrappers.R
\name{getInfoByEmail}
\alias{getInfoByEmail}
\title{Get Maintainer Information by Maintainer Email}
\usage{
getInfoByEmail(email)
}
\arguments{
\item{email}{current Bioconductor package maintainer email}
}
\value{
data.frame of maintainer information including name, package, email,
consent_date, email_status, is_email_valid
}
\description{
Retrieves maintainer information by maintainer email
}
\details{
Retrieves maintainer information by maintainer email
}
\examples{
  tbl <- getInfoByEmail("maintainer@bioconductor.org")

}
\author{
Lori Shepherd
}
