% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiWrappers.R
\name{getInfoByName}
\alias{getInfoByName}
\title{Get Maintainer Information by Maintainer Name}
\usage{
getInfoByName(name)
}
\arguments{
\item{name}{current Bioconductor package maintainer name}
}
\value{
data.frame of maintainer information including name, package, email,
consent_date, email_status, is_email_valid
}
\description{
Retrieves maintainer information by maintainer name
}
\details{
Retrieves maintainer information by maintainer name
}
\examples{
  tbl <- getInfoByName("Lori Shepherd")
  tbl <- getInfoByName("Hervé Pagès")

}
\author{
Lori Shepherd
}
