% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiWrappers.R
\name{getInfoByPackage}
\alias{getInfoByPackage}
\title{Get Maintainer Information by Package Name}
\usage{
getInfoByPackage(packageName)
}
\arguments{
\item{packageName}{current Bioconductor package}
}
\value{
data.frame of maintainer information including name, package, email,
consent_date, email_status, is_email_valid
}
\description{
Retrieves maintainer information by package name
}
\details{
Retrieves maintainer information by package name
}
\examples{
  tbl <- getInfoByPackage("BiocFileCache")

}
\author{
Lori Shepherd
}
