% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiWrappers.R
\name{isEmailValid}
\alias{isEmailValid}
\title{Check if Maintainer Email is Valid}
\usage{
isEmailValid(email)
}
\arguments{
\item{email}{current Bioconductor package maintainer email}
}
\value{
list. All lists have attribute 'valid' which is TRUE/FALSE. If valid
is FALSE, the list will also contain an attribute 'data' with additional
diagnostic columns in the database. This includes: email, name, package,
consent_date, email_status, is_email_valid, bounce_type, bounce_subtype,
smtp_status, diagnostic_code. If no 'data' attribute is included, the email
is not found in the database.
}
\description{
Retrieves if maintainer email is valid
}
\details{
Retrieves if maintainer email is valid.
}
\examples{
  tbl <- getInfoByEmail("maintainer@bioconductor.org")

}
\author{
Lori Shepherd
}
