% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiWrappers.R
\name{listEmailsOnSuppressionList}
\alias{listEmailsOnSuppressionList}
\title{List All Emails on Suppression List}
\usage{
listEmailsOnSuppressionList()
}
\value{
list. First attribute is 'awssuppression' which is TRUE. Second
attribute is 'data' which is a data.frame that includes the email and name.
}
\description{
Retrieves all emails on AWS Suppression List
}
\details{
Retrieves all emails on AWS Suppression List
}
\examples{
  tbl <- listEmailsOnSuppressionList()

}
\author{
Lori Shepherd
}
