% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveTable.R
\name{get_maintainer_data}
\alias{get_maintainer_data}
\title{Retrieve Maintainer Information}
\usage{
get_maintainer_data(
  url = "https://pkgmaintainers.bioconductor.org/download-maintainer-db"
)
}
\arguments{
\item{url}{Url to retrieve data from whose data can be parsed with jsonlite::fromJSON}
}
\value{
data.frame of maintainer information
}
\description{
Retrieves maintainer information from the package maintainer
validation app
}
\details{
Retrieves maintainer information from the package maintainer
validation app
}
\examples{
  tbl <- get_maintainer_data()

}
\author{
Lori Shepherd
}
