% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiWrappers.R
\name{listAllBadEmails}
\alias{listAllBadEmails}
\title{List All 'Bad' Emails}
\usage{
listAllBadEmails()
}
\value{
single column matrix that lists all maintainer emails.
}
\description{
Retrieves all 'bad' emails.
}
\details{
Retrieves all 'bad' emails.
}
\examples{
  tbl <- listAllBadEmails()

}
\author{
Lori Shepherd
}
