% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiWrappers.R
\name{listInvalid}
\alias{listInvalid}
\title{List All Invalid Emails}
\usage{
listInvalid()
}
\value{
data.frame of maintainer information for anyone with a currently
identified invalid email. This includes: email, name, package,
email_status, is_email_valid, bounce_type, bounce_subtype, smtp_status,
diagnostic_code. An email is designated invalid if the validation email
failed to send either for a bounce or because it is on the AWS suppression
list.
}
\description{
Retrieves all invalid emails
}
\details{
Retrieves all invalid emails
}
\examples{
  tbl <- listInvalid()

}
\author{
Lori Shepherd
}
