% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiWrappers.R
\name{listNeedsConsent}
\alias{listNeedsConsent}
\title{List All Maintainer that Need Consent}
\usage{
listNeedsConsent()
}
\value{
character vector that lists all maintainer emails that need consent.
}
\description{
Retrieves all maintainers that need to consent to Bioconductor policies.
}
\details{
Retrieves all maintainers that need to consent to Bioconductor
policies. Bioconductor requires maintainers consent to Bioconductor policies
once a year.
}
\examples{
  tbl <- listNeedsConsent()

}
\author{
Lori Shepherd
}
